requirements.txt
asgiref==3.8.1
certifi==2024.8.30
channels==4.2.0
charset-normalizer==3.4.0
Django==5.1.4
djangorestframework==3.15.2
dnspython==2.7.0
docker==7.1.0
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
mysqlclient==2.2.6
numpy==2.2.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
psycopg2-binary==2.9.10
pymongo==4.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.2.3
